+++ example/main.dart
@@ -4,7 +4,7 @@ import 'package:dartcore/dartcore.dart' as dartcore;
 void main() async {
   var app = dartcore.App();
 
-  // Custom 404 Handler
+  // custom 404 error
   app.set404((request) {
     request.response
       ..statusCode = HttpStatus.notFound
@@ -12,7 +12,7 @@ void main() async {
       ..close();
   });
 
-  // Custom 500 Handler
+  // custom 500 error
   app.set500((request, error) {
     request.response
       ..statusCode = HttpStatus.internalServerError
@@ -20,27 +20,27 @@ void main() async {
       ..close();
   });
 
-  // Middleware for logging
+  // middleware for logging
   app.use((request, next) async {
     print('[Middleware] ${request.method} ${request.uri}');
     await next();
   });
 
-  // Route for serving static files
+  // serving static files
   app.route('GET', '/static/<file>', (request) async {
     var filePath = request.uri.pathSegments[2];
     await app.serveStaticFile(request, 'static/$filePath');
   });
 
-  // Route for handling JSON POST requests
+  // JSON POST requests
   app.route('POST', '/json', (request) async {
     var jsonData = await app.parseJson(request);
     await app.sendJson(request, {'received': jsonData});
   });
 
-  // Route for handling file uploads
+  // file uploads      -- Make the directory "uploads" before executing, else the server will crash with an OS error.
   app.route('POST', '/upload', (request) async {
-    await app.parseMultipartRequest(request);
+    await app.parseMultipartRequest(request, 'uploads');
     request.response
       ..statusCode = HttpStatus.ok
       ..write('File uploaded successfully.\n')