+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
 - Initial version.
+
+## 0.0.2
+
+- Added Rate limit support
+- Added CORS support
+- Added Custom 404 and 500 responses
+- Added 0.0.0.0 support
+- Added Rate limit support (Uses Memory for now)
+- Added multipart support